chore: avoid using 'if runtime.GOOS'
