+++ edge/cmd/edgecore/app/server.go
@@ -176,6 +176,7 @@ func environmentCheck() error {
 	for _, process := range processes {
 		processName, err := process.Name()
 		if err != nil {
+			// err usually cant be fixed by user, so just skip it instead of return error (we only check named process)
 			continue
 		}
 		switch processName {
