+++ edge/pkg/devicetwin/dmiserver/server.go
@@ -21,8 +21,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"net"
-	"os"
-	"runtime"
 	"sync"
 	"time"
 
@@ -177,32 +175,6 @@ func CreateMessageTwinUpdate(name, valueType, value string) ([]byte, error) {
 
 	msg, err := json.Marshal(updateMsg)
 	return msg, err
-}
-
-func initSock(sockPath string) error {
-	klog.Infof("init uds socket: %s", sockPath)
-	if runtime.GOOS == "windows" {
-		err := os.Remove(sockPath)
-		if os.IsNotExist(err) {
-			return nil
-		} else if err != nil {
-			klog.Error(err)
-			return fmt.Errorf("fail to stat uds socket path")
-		}
-		return nil
-	}
-	_, err := os.Stat(sockPath)
-	if err == nil {
-		err = os.Remove(sockPath)
-		if err != nil {
-			return err
-		}
-		return nil
-	} else if os.IsNotExist(err) {
-		return nil
-	} else {
-		return fmt.Errorf("fail to stat uds socket path")
-	}
 }
 
 func StartDMIServer(cache *DMICache) {
